pyproject.toml
[tool.poetry]
name = "poetry-onlineshop"
version = "0.1.0"
description = ""
authors = ["Mantile <ggwp1337@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-unfold = "^0.39.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
